nexus-backend/nexus/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
langgraph==0.2.28
langchain-core==0.3.15
groq==1.1.2
httpx==0.28.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-docx==1.1.2
dateparser==1.2.0
python-dotenv==1.0.1
python-multipart==0.0.12
sentence-transformers==2.7.0
scikit-learn==1.5.2
numpy>=1.24
pandas>=2.0
datasets>=2.18.0
pypdf>=4.0.0
transformers>=4.40.0
supabase==2.10.0
